Vulnerabilities/2020/CVE-2020-25784/tezeb/requirements.txt
bcrypt==3.2.0
capstone==4.0.2
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
idna==2.10
intervaltree==3.1.0
Mako==1.1.3
MarkupSafe==1.1.1
packaging==20.8
paramiko==2.7.2
psutil==5.8.0
pwntools==4.3.1
pycparser==2.20
pyelftools==0.27
Pygments==2.7.3
PyNaCl==1.4.0
pyparsing==2.4.7
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.1
requests==2.25.1
ROPGadget==6.4
six==1.15.0
sortedcontainers==2.3.0
unicorn==1.0.2rc3
urllib3==1.26.2
